Only make available actions that don't have `is_complete` set to True
This addresses the second point in issue #39 ("Reviewers can take their action
multiple times"). The first issue ("Multiple assignments for a reviewer") should
not be an issue because of commit de6b8cf47e01422b0911496efb9f614032b00a45, in
which only assignments with `is_complete` set to False are taken into account for
the awaiting query.

Also, the route returns cases, not assignments (and since we removed the
assignment token from the serialized case in commit
60b35efb41a9a63fafd4b57f3506fba24d417df2, this should be less ambiguous). So this
commit should close #39.